+++ src/LazyCart/ILazyCartesianProduct.cs
@@ -10,6 +10,11 @@ namespace LazyCart;
 public interface ILazyCartesianProduct<T1, T2>
 {
     /// <summary>
+    ///     Gets the size of the Cartesian product.
+    /// </summary>
+    BigInteger Size { get; }
+
+    /// <summary>
     ///     Returns the Cartesian product entry at the given index.
     /// </summary>
     /// <param name="index">The index of the Cartesian product entry to return.</param>
@@ -32,4 +37,12 @@ public interface ILazyCartesianProduct<T1, T2>
     /// <exception cref="ArgumentOutOfRangeException">Thrown if the Cartesian product entry is not found.</exception>
     /// <returns>The index of the given Cartesian product entry.</returns>
     BigInteger IndexOf((T1, T2) entry);
+
+    /// <summary>
+    ///     Generates an evenly-distributed, random sample of Cartesian product entries.
+    /// </summary>
+    /// <param name="sampleSize">The number of samples to generate.</param>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown if <paramref name="sampleSize"/> is greater than <see cref="Size"/>.</exception>
+    /// <returns>An evenly-distributed, random sample of Cartesian product entries.</returns>
+    IEnumerable<(T1, T2)> GenerateSamples(BigInteger sampleSize);
 }