+++ tests/LazyCart.Tests/LazyCartesianProductPropertyTests.cs
@@ -13,8 +13,8 @@ public sealed class LazyCartesianProductPropertyTests
 
     public LazyCartesianProductPropertyTests()
     {
-        _set1 = Enumerable.Range(1, 100000).ToList();
-        _set2 = Enumerable.Range(1, 100000).Select(i => i.ToString(CultureInfo.CurrentCulture)).ToList();
+        _set1 = Enumerable.Range(1, 500000).ToList();
+        _set2 = Enumerable.Range(1, 500000).Select(i => i.ToString(CultureInfo.CurrentCulture)).ToList();
         _subjectUnderTest = new LazyCartesianProduct<int, string>(_set1, _set2);
     }
 
@@ -49,5 +49,23 @@ public sealed class LazyCartesianProductPropertyTests
         var result = _subjectUnderTest!.IndexOf(entry);
 
         result.Should().BeInRange(0, _set1!.Count * _set2!.Count - 1);
+    }
+
+    [Property(MaxTest = 100000)]
+    public void GenerateSamples_ShouldReturnDistinctValues(NonNegativeInt sampleSize)
+    {
+        if (sampleSize.Get < _set1!.Count * _set2!.Count)
+        {
+            var samples = _subjectUnderTest!.GenerateSamples(sampleSize.Get).ToList();
+
+            samples.Should().HaveCount(sampleSize.Get);
+            samples.Should().OnlyHaveUniqueItems();
+        }
+        else
+        {
+            var act = () => _subjectUnderTest!.GenerateSamples(sampleSize.Get);
+
+            act.Should().Throw<ArgumentOutOfRangeException>();
+        }
     }
 }