+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] 2023-05-14
+
+### Added
+
+New methods for getting the Cartesian product size as well as generating a random sample
+of Cartesian product entries.
+
 ## [0.1.3] 2023-05-14
 
 ### Added